requirements.txt
torch
torchvision
Pillow
streamlit
gdown
numpy
